+++ pooch/core.py
@@ -18,13 +18,7 @@ from .utils import (
     hash_algorithm,
     hash_matches,
 )
-from .downloaders import HTTPDownloader, FTPDownloader
-
-KNOWN_DOWNLOADERS = {
-    "ftp": FTPDownloader,
-    "https": HTTPDownloader,
-    "http": HTTPDownloader,
-}
+from .downloaders import choose_downloader
 
 
 def create(
@@ -365,16 +359,9 @@ class Pooch:
                 str(self.path),
             )
 
-            parsed_url = parse_url(url)
-            if parsed_url["protocol"] not in KNOWN_DOWNLOADERS:
-                raise ValueError(
-                    "Unrecognized URL protocol '{}' in '{}'. Must be one of {}.".format(
-                        parsed_url["protocol"], url, KNOWN_DOWNLOADERS.keys()
-                    )
-                )
-
             if downloader is None:
-                downloader = KNOWN_DOWNLOADERS[parsed_url["protocol"]]()
+                downloader = choose_downloader(url)
+
             # Stream the file to a temporary so that we can safely check its
             # hash before overwriting the original
             tmp = tempfile.NamedTemporaryFile(delete=False, dir=str(self.abspath))
